# perf: Faster CLI