--- a/bench/app.py
+++ b/bench/app.py
@@ -11,7 +11,6 @@ import sys
 
 # imports - third party imports
 import click
-from six.moves import reload_module
 
 # imports - module imports
 import bench
@@ -364,6 +363,7 @@ def get_repo_dir(app, bench_path='.'):
 
 def switch_branch(branch, apps=None, bench_path='.', upgrade=False, check_upgrade=True):
 	import git
+	from six.moves import reload_module
 	from bench.utils import update_requirements, update_node_packages, backup_all_sites, patch_sites, build_assets, post_upgrade
 	apps_dir = os.path.join(bench_path, 'apps')
 	version_upgrade = (False,)