--- a/bench/config/__init__.py
+++ b/bench/config/__init__.py
@@ -1,6 +1,5 @@
 """Module for setting up system and respective bench configurations"""
 
-# imports - third party imports
 
 def env():
 	from jinja2 import Environment, PackageLoader