--- a/bench/config/common_site_config.py
+++ b/bench/config/common_site_config.py
@@ -3,7 +3,6 @@ import getpass
 import json
 import os
 
-# imports - third party imports
 
 
 default_config = {