--- a/bench/config/redis.py
+++ b/bench/config/redis.py
@@ -5,7 +5,6 @@ import subprocess
 
 # imports - third party imports
 import semantic_version
-from six.moves.urllib.parse import urlparse
 
 # imports - module imports
 import bench
@@ -13,6 +12,8 @@ from bench.config.common_site_config import get_config
 
 
 def generate_config(bench_path):
+	from six.moves.urllib.parse import urlparse
+
 	config = get_config(bench_path)
 
 	ports = {}