--- a/bench/config/supervisor.py
+++ b/bench/config/supervisor.py
@@ -22,7 +22,7 @@ def generate_supervisor_config(bench_path, user=None, yes=False):
 	if not user:
 		user = getpass.getuser()
 
-	template = bench.config.env.get_template('supervisor.conf')
+	template = bench.config.env().get_template('supervisor.conf')
 	config = get_config(bench_path=bench_path)
 	bench_dir = os.path.abspath(bench_path)