--- a/bench/utils.py
+++ b/bench/utils.py
@@ -423,7 +423,7 @@ def setup_sudoers(user):
 		if set_permissions:
 			os.chmod('/etc/sudoers', 0o440)
 
-	template = bench.config.env.get_template('frappe_sudoers')
+	template = bench.config.env().get_template('frappe_sudoers')
 	frappe_sudoers = template.render(**{
 		'user': user,
 		'service': find_executable('service'),