This doesn't look tasteful, but at least it's fast. Well, I'll let you decide if it's fast. This is how long a typical bench CLI invocation takes.


```
$ time bench --site speed.com execute frappe.ping  
0.69s user 0.09s system 100% cpu 0.770 total
```
We can improve this by not importing code that isn't executed e.g. Jinja, Requests, etc, and shed ~300ms.
```
$ time bench --site speed.com execute frappe.ping  
0.28s user 0.06s system 101% cpu 0.329 total
```
### Profiling
Python provides a non-invasive way to profile import times with [`-X importtime`](https://docs.python.org/3/using/cmdline.html#id5). Executing `bench` doesn't seem to be compatible with this flag but, we can create `main.py`

```python3
from bench.cli import cli

if __name__ == "__main__":
	cli()
```
and execute
```
python -X importtime ../../bench/bench/main.py --site speed.com execute frappe.ping
```
**Note:** `../../bench/bench/main.py` is the path where I had placed the main.py file.

```
import time: self [us] | cumulative | imported package
import time:       102 |        102 | zipimport
import time:       636 |        636 | _frozen_importlib_external
...
...
import time:       382 |        382 |     bench.commands.install
import time:       402 |       5915 |   bench.commands
import time:       308 |      36098 | bench.cli
"pong"
```
We can then use [tuna](https://github.com/nschloe/tuna) for visualization with something like.
```
python -X importtime ../../bench/bench/main.py --site speed.com execute frappe.ping 2> import.log
tuna import.log
```
### Screenshots
**Before**
![Screenshot_2021-02-19 tuna - import log(15)](https://user-images.githubusercontent.com/8528887/108661229-552ec180-74f1-11eb-8ae0-2357cd12c6bc.png)
**After**
![Screenshot_2021-02-22 tuna - import log](https://user-images.githubusercontent.com/8528887/108661260-65df3780-74f1-11eb-81fd-a5de1441399c.png)

Actually, the CLI is slower than shown here, https://github.com/frappe/frappe/pull/12447 has a similar effect on Frappe.
Before Both PRs
```
$ time bench --site speed.com execute frappe.ping  
1.03s user 0.13s system 100% cpu 1.152 total
```
After this PR (and before frappe PR)
```
$ time bench --site speed.com execute frappe.ping  
0.69s user 0.09s system 100% cpu 0.770 total
```
After both PRs
```
$ time bench --site speed.com execute frappe.ping  
0.28s user 0.06s system 101% cpu 0.329 total
```
**Note:** If you're using editable install then you might see worse numbers (~200ms slower), if not then you might have to reinstall `bench` after every change.