@Alchez 2021-02-23
@adityahase 

Just curious, are the benchmarks just for the `ping` function?

Wouldn't function imports lead to a slower call to the specific command, like `get-app` or `install-app`? What do the benchmarks look like for those commands?