@gavindsouza 2021-02-26
> Wouldn't function imports lead to a slower call to the specific command, like `get-app` or `install-app`? What do the benchmarks look like for those commands?

@Alchez I don't think these changes increase execution times for any particular command as the same modules would be imported in any way that are required by a certain command. So, best case scenario, 
`bench:cli` hits one of the functions this PR touches and only the required modules needs to be imported, which would make execution faster. On the other hand, for a function that requires all the modules to be imported won't suffer from degradation of performance as it would take the Python virtual machine the same amount of time to import the same set of modules in whatever order.

But I'm eager to see benchmarks, just out of curiosity. @adityahase do me let me know if I've missed or misunderstood anything here.

---

Anyhoo, merging this 👍🏼